Move Backend enum to backends/mod.rs
