+++ src/backends/mod.rs
@@ -1,5 +1,10 @@
 use log::error;
 
+use crate::backends::{
+	udev::UdevData,
+	winit::WinitData,
+};
+
 // Copyright 2023 the Strata authors
 // SPDX-License-Identifier: GPL-3.0-or-later
 
@@ -16,6 +21,11 @@ pub mod cursor;
 mod drawing;
 pub mod udev;
 pub mod winit;
+
+pub enum Backend {
+	Winit(WinitData),
+	Udev(UdevData),
+}
 
 pub enum Backend {
 	Winit(WinitData),
