+++ src/state.rs
@@ -102,11 +102,7 @@ use smithay::{
 };
 
 use crate::{
-	backends::{
-		udev::UdevData,
-		winit::WinitData,
-		Backend,
-	},
+	backends::Backend,
 	decorations::BorderShader,
 	handlers::input::{
 		KeyPattern,
@@ -252,29 +248,6 @@ impl StrataState {
 
 		Ok(())
 	}
-}
-
-pub struct WinitData {
-	backend: WinitGraphicsBackend<GlowRenderer>,
-	damage_tracker: OutputDamageTracker,
-}
-
-pub struct UdevData {
-	pub session: LibSeatSession,
-	dh: DisplayHandle,
-	dmabuf_state: Option<(DmabufState, DmabufGlobal)>,
-	primary_gpu: DrmNode,
-	allocator: Option<Box<dyn Allocator<Buffer = Dmabuf, Error = AnyError>>>,
-	gpus: GpuManager<GbmGlesBackend<GlesRenderer>>,
-	backends: HashMap<DrmNode, BackendData>,
-	pointer_images: Vec<(xcursor::parser::Image, TextureBuffer<MultiTexture>)>,
-	pointer_element: PointerElement<MultiTexture>,
-	pointer_image: crate::backends::cursor::Cursor,
-}
-
-pub enum Backend {
-	Winit(WinitData),
-	Udev(UdevData),
 }
 
 pub struct StrataComp {
